ДЗ4 - Добавляю решение v1.01
Были добавлены проверки для кода и сообщения с статусом выполнения операции в print()

+++ ДЗ 4 - Работа с PostgreSQL из Python/main.py
@@ -49,7 +49,26 @@ def add_client(conn, name: str, surname: str, email: str, phone=None) -> int:
 
         Returns:
             int: The ID of the newly added client.
+
+        Raises:
+            TypeError: If the name, surname, or email data type is not str.
+            ValueError: If name, surname or email is blank or None.
     """
+    if not isinstance(name, str):
+        raise TypeError('name must be a string')
+    if not isinstance(surname, str):
+        raise TypeError('surname should be a string')
+    if not isinstance(email, str):
+        raise TypeError('email should be a string')
+    if not name.strip():
+        raise ValueError('name cannot be empty')
+    if not surname.strip():
+        raise ValueError('surname cannot be empty')
+    if not email.strip():
+        raise ValueError('email cannot be empty')
+    if phone is not None and not isinstance(phone, str):
+        raise TypeError('phone must be a string or None')
+
     with conn.cursor() as cur:
         cur.execute(
             """
@@ -57,13 +76,15 @@ def add_client(conn, name: str, surname: str, email: str, phone=None) -> int:
             VALUES (%s, %s, %s)
             RETURNING id
             """,
-            (name, surname, email)
+            (name.strip(), surname.strip(), email.strip())
         )
         client_id = cur.fetchone()[0]
-        
+
         if phone:
             add_phone(conn, client_id, phone)
-        
+
+        print(f'Client {name} {surname} added successfully! Client ID: {client_id}')
+        print(f'Email: {email}. Phone: {phone}', end='\n\n')
         conn.commit()
         return client_id
 
@@ -79,16 +100,34 @@ def add_phone(conn, client_id: int, phone: str) -> None:
 
         Returns:
             None
+
+        Raises:
+            TypeError: If client_id is not an integer or phone is not a string.
+            ValueError: If client_id is None or phone is an empty string.
     """
+    if not isinstance(client_id, int):
+        raise TypeError('client_id must be an integer')
+    if not isinstance(phone, str):
+        raise TypeError('phone should be a string')
+    if not phone.strip():
+        raise ValueError("phone can't be empty")
+    if client_id is None:
+        raise ValueError('client_id cannot be None')
+
     with conn.cursor() as cur:
         cur.execute(
             """
             INSERT INTO phones (client_id, phone)
             VALUES (%s, %s)
+            RETURNING id
             """,
             (client_id, phone)
         )
+
+        phone_id = cur.fetchone()[0]
         conn.commit()
+        print(f"Phone '{phone}', with ID {phone_id}, has been successfully added"
+              f"to the client's account with ID: {client_id}.", end='\n\n')
 
 
 def change_client_info(
@@ -114,17 +153,17 @@ def change_client_info(
         updates, values = [], []
 
         if name is not None:
-            updates.append("name = %s")
+            updates.append('name = %s')
             values.append(name)
         if surname is not None:
-            updates.append("surname = %s")
+            updates.append('surname = %s')
             values.append(surname)
         if email is not None:
-            updates.append("email = %s")
+            updates.append('email = %s')
             values.append(email)
 
         if updates:
-            query = "UPDATE clients SET " + ", ".join(updates) + " WHERE id = %s"
+            query = 'UPDATE clients SET ' + ', '.join(updates) + ' WHERE id = %s'
             values.append(client_id)
             cur.execute(query, values)
 
@@ -138,6 +177,7 @@ def change_client_info(
             )
 
         conn.commit()
+        print(f"The client's data with ID {client_id} has been successfully updated!", end='\n\n')
 
 
 def del_phone(conn, client_id: int, phone: str) -> None:
@@ -151,17 +191,27 @@ def del_phone(conn, client_id: int, phone: str) -> None:
 
         Returns:
             None
+
+        Raises:
+            TypeError: If client_id is not an integer or phone is not a string.
+            ValueError: If either client_id or phone is None or empty.
     """
-    if not client_id or not phone:
-        return None
+    if not isinstance(client_id, int):
+        raise TypeError('client_id must be an integer')
+    if not isinstance(phone, str):
+        raise TypeError('phone must be a string')
+    if not client_id or not phone.strip():
+        raise ValueError('Client ID and phone number must be provided and not empty')
 
     phone_id = get_phone_id(conn, client_id, phone)
     with conn.cursor() as cur:
         cur.execute('DELETE FROM phones WHERE id = %s', (phone_id,))
         conn.commit()
+        print(f"Phone '{phone}' has been successfully deleted"
+              f"from the client's account with ID: {client_id}.", end='\n\n')
 
 
-def get_phone_id(conn, client_id: int, phone: str) -> int or None:
+def get_phone_id(conn, client_id: int, phone: str) -> int | None:
     """
         Returns the ID of the phone number in the database.
 
@@ -172,13 +222,18 @@ def get_phone_id(conn, client_id: int, phone: str) -> int or None:
 
         Returns:
             int: The ID of the phone number in the database.
+            None: If the phone number is not found in the database.
+
+        Raises:
+            ValueError: If either client_id or phone is None.
     """
     if not client_id or not phone:
-        return None
+        raise ValueError('Client ID and phone number must be provided')
 
     with conn.cursor() as cur:
         cur.execute('SELECT id FROM phones WHERE client_id = %s AND phone = %s', (client_id, phone))
-        return cur.fetchone()[0]
+        result = cur.fetchone()
+        return result[0] if result else None
 
 
 def del_client(conn, client_id: int) -> None:
@@ -191,16 +246,30 @@ def del_client(conn, client_id: int) -> None:
 
         Returns:
             None
+
+        Raises:
+            TypeError: If client_id is not an integer.
+            ValueError: If client_id does not exist in the database.
     """
+    if not isinstance(client_id, int):
+        raise TypeError('Client ID must be an integer')
+
     with conn.cursor() as cur:
+        cur.execute('SELECT id FROM clients WHERE id = %s', (client_id,))
+        client_exists = cur.fetchone() is not None
+
+        if not client_exists:
+            raise ValueError('Client ID does not exist in the database')
+
         cur.execute('DELETE FROM clients WHERE id = %s', (client_id,))
         conn.commit()
+        print(f'The client with the identifier {client_id} has been permanently deleted!', end='\n\n')
 
 
 def find_client(
         conn, name=None, surname=None,
         email=None, phone=None
-) -> list:
+) -> list | str:
     """
         Finds clients in the database that match the given criteria.
 
@@ -212,8 +281,25 @@ def find_client(
             phone (str): The phone number of the client.
 
         Returns:
-            list: A list of client IDs that match the given criteria.
+            list: A list of client IDs that match the given criteria in the database.
+            str: A message indicating that no clients were found that match the given criteria.
+
+        Raises:
+            TypeError: If name, surname, email, or phone are not strings or None.
+            ValueError: If all search criteria (name, surname, email, phone) are None.
     """
+    if name is not None and not isinstance(name, str):
+        raise TypeError('name must be a string or Non')
+    if surname is not None and not isinstance(surname, str):
+        raise TypeError('surname must be a string or None')
+    if email is not None and not isinstance(email, str):
+        raise TypeError('email must be a string or None')
+    if phone is not None and not isinstance(phone, str):
+        raise TypeError('phone must be a string or None')
+
+    if all(arg is None for arg in [name, surname, email, phone]):
+        raise ValueError('At least one search criterion must be provided')
+
     with conn.cursor() as cur:
         query = """
             SELECT id FROM clients
@@ -224,7 +310,11 @@ def find_client(
         """
         params = [name, name, surname, surname, email, email, phone, phone]
         cur.execute(query, params)
-        return [row[0] for row in cur.fetchall()]
+        if cur.rowcount != 0:
+            print(f'The unique identifier of the identified client: ')
+            return [row[0] for row in cur.fetchall()]
+        else:
+            return 'Warning: The search for the specified criteria was unsuccessful'
 
 
 with psycopg2.connect(
@@ -234,18 +324,26 @@ with psycopg2.connect(
     database=database
 ) as conn:
     create_db(conn)
-    # chanmina = add_client(conn, 'Chanmina', 'Chanmina', 'chanmina@example.com')
-    # add_phone(conn, chanmina, '+193459463')
-    # din = add_client(conn, 'Din', 'Sakay', 'dinsakay@example.com', '+193245463')
-    # add_phone(conn, 18, '+144245463')
-    # change_client_info(
-    #     conn, client_id=13, name='Emely',
-    #     surname='James', email='emelyjames@example.com', phones=['+193446464']
-    # )
-    # del_phone(conn, 20)
-    # del_client(conn, 18)
-    print(find_client(conn, 'John'))
-    print(find_client(conn, email='johndoe12@example.com'))
+
+    # Adding some clients
+    # add_client(conn, 'John', 'Doe', 'jdoe@ex.com')
+
+    # Adding some phones
+    # add_phone(conn, 23, '+1234567890')
+    # add_phone(conn, 23, '+1234567256')
+
+    # Deleting some phones
+    # del_phone(conn, 23, '+1234567890')
+    # del_phone(conn, 23, '+1234567890')
+
+    # Deleting some clients
+    # del_client(conn, 23)
+
+    # Updating
+    # change_client_info(conn, 23, email='mikeanodas@ex.com')
+
+    # Finding clients
+    # print(find_client(conn, name='Doe'), '\n')
 
     # Checking results from queries above in the terminal
     with conn.cursor() as cur:
